Add InstanceStatusRegistryService
