+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/GovernanceFacade.java
@@ -79,7 +79,7 @@ public final class GovernanceFacade implements AutoCloseable {
      * Online instance.
      */
     public void onlineInstance() {
-        registryCenter.registerInstanceOnline();
+        registryCenter.getInstanceStatusService().registerInstanceOnline(registryCenter.getInstanceId());
         listenerFactory.watchListeners();
     }
