+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/registry/RegistryCenter.java
@@ -30,8 +30,8 @@ import org.apache.shardingsphere.governance.core.registry.config.subscriber.Glob
 import org.apache.shardingsphere.governance.core.registry.config.subscriber.SchemaRuleRegistrySubscriber;
 import org.apache.shardingsphere.governance.core.registry.metadata.service.SchemaRegistryService;
 import org.apache.shardingsphere.governance.core.registry.process.subscriber.ProcessRegistrySubscriber;
-import org.apache.shardingsphere.governance.core.registry.state.node.StatesNode;
 import org.apache.shardingsphere.governance.core.registry.state.service.DataSourceStatusRegistryService;
+import org.apache.shardingsphere.governance.core.registry.state.service.InstanceStatusRegistryService;
 import org.apache.shardingsphere.governance.core.registry.state.service.UserStatusRegistryService;
 import org.apache.shardingsphere.governance.core.registry.state.subscriber.DataSourceStatusRegistrySubscriber;
 import org.apache.shardingsphere.governance.repository.spi.RegistryCenterRepository;
@@ -47,42 +47,36 @@ import java.util.Properties;
 /**
  * Registry center.
  */
+@Getter
 public final class RegistryCenter {
     
     private final String instanceId;
     
-    private final RegistryCenterRepository repository;
-    
-    @Getter
     private final DataSourcePersistService dataSourceService;
     
-    @Getter
     private final SchemaRulePersistService schemaRuleService;
     
-    @Getter
     private final GlobalRulePersistService globalRuleService;
     
-    @Getter
     private final PropertiesPersistService propsService;
     
-    @Getter
     private final SchemaRegistryService schemaService;
     
-    @Getter
     private final DataSourceStatusRegistryService dataSourceStatusService;
     
-    @Getter
+    private final InstanceStatusRegistryService instanceStatusService;
+    
     private final LockRegistryService lockService;
     
     public RegistryCenter(final RegistryCenterRepository repository) {
         instanceId = GovernanceInstance.getInstance().getId();
-        this.repository = repository;
         dataSourceService = new DataSourcePersistService(repository);
         schemaRuleService = new SchemaRulePersistService(repository);
         globalRuleService = new GlobalRulePersistService(repository);
         propsService = new PropertiesPersistService(repository);
         schemaService = new SchemaRegistryService(repository);
         dataSourceStatusService = new DataSourceStatusRegistryService(repository);
+        instanceStatusService = new InstanceStatusRegistryService(repository);
         lockService = new LockRegistryService(repository);
         createSubscribers(repository);
     }
@@ -115,14 +109,5 @@ public final class RegistryCenter {
             dataSourceService.persist(schemaName, dataSourceConfigs.get(schemaName), isOverwrite);
             schemaRuleService.persist(schemaName, schemaRuleConfigs.get(schemaName), isOverwrite);
         }
-    }
-    
-    /**
-     * Register instance online.
-     */
-    public void registerInstanceOnline() {
-        repository.persist(StatesNode.getDataNodesPath(), "");
-        repository.persist(StatesNode.getPrimaryNodesPath(), "");
-        repository.persistEphemeral(StatesNode.getProxyNodePath(instanceId), "");
     }
 }
