+++ shardingsphere-governance/shardingsphere-governance-core/src/test/java/org/apache/shardingsphere/governance/core/GovernanceFacadeTest.java
@@ -40,6 +40,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertThat;
+import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
@@ -65,7 +66,7 @@ public final class GovernanceFacadeTest {
     
     @Test
     public void assertOnlineInstance() {
-        RegistryCenter registryCenter = mock(RegistryCenter.class);
+        RegistryCenter registryCenter = mock(RegistryCenter.class, RETURNS_DEEP_STUBS);
         GovernanceWatcherFactory listenerFactory = mock(GovernanceWatcherFactory.class);
         setField(governanceFacade, "registryCenter", registryCenter);
         setField(governanceFacade, "listenerFactory", listenerFactory);
@@ -75,7 +76,6 @@ public final class GovernanceFacadeTest {
         Properties props = new Properties();
         governanceFacade.onlineInstance(Collections.singletonMap("sharding_db", dataSourceConfigs), schemaRuleConfigs, globalRuleConfigs, props);
         verify(registryCenter).persistConfigurations(Collections.singletonMap("sharding_db", dataSourceConfigs), schemaRuleConfigs, globalRuleConfigs, props, false);
-        verify(registryCenter).registerInstanceOnline();
         verify(listenerFactory).watchListeners();
     }
